add loader when loading repos to oss page
