+++ packages/ui/src/components/LogoLoader/index.ts
@@ -1,0 +1,3 @@
+import LogoLoader from './LogoLoader'
+
+export default LogoLoader
